task: Follow conventions previously used
